services/alarm-receiver/requirements.txt
aiohttp==3.9.3
python-dotenv==1.0.0
aiosqlite==0.20.0
